pyproject.toml
[tool.poetry]
name = "ether_transaction"
version = "0.1.0"
description = ""
authors = ["Nikita Maslov <maslov.n.e@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^5.27.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
